api/requirements.txt
FastAPI
psycopg2-binary
uvicorn
